requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.0.0
tqdm>=4.60.0
thop>=0.1.1
kagglehub>=0.3.0
